added custom installation for this config
There is this 'setup.sh' that will install the configuration of this
repository automatically.

It will check the type architecture and install depending the needs of a
Nix with packages and home-manager or a full NixOS environment.

It also supports arm64 with Nix packages and home-manager.

Details about it are in the files comments and the READMEs.